photon-order-error event
