+++ packages/elements/src/photon-multirx-form/components/AddPrescriptionCard.tsx
@@ -8,9 +8,9 @@ import {
   ScreeningAlertType,
   Text,
   triggerToast,
-  usePrescribe
+  usePrescribe,
+  usePrescribeEventDispatch
 } from '@photonhealth/components';
-import photonStyles from '@photonhealth/components/dist/style.css?inline';
 import { DispenseUnit, Medication, Prescription } from '@photonhealth/sdk/dist/types';
 import { any, min, number, optional, record, refine, size, string } from 'superstruct';
 import { afterDate, between, message } from '../../validators';
@@ -62,12 +62,12 @@ export const AddPrescriptionCard = (props: {
   disableList?: DisableList;
 }) => {
   const { tryCreatePrescription } = usePrescribe();
+  const { dispatchOrderError } = usePrescribeEventDispatch();
   const [offCatalog, setOffCatalog] = createSignal<Medication | undefined>(undefined);
   const [dispenseUnit] = createSignal<DispenseUnit | undefined>(undefined);
   const [openDoseCalculator, setOpenDoseCalculator] = createSignal(false);
   const [searchText, setSearchText] = createSignal<string>('');
   const [isLoading, setIsLoading] = createSignal(false);
-  let ref: any;
 
   onMount(() => {
     for (const [k, v] of Object.entries(validators)) {
@@ -80,17 +80,6 @@ export const AddPrescriptionCard = (props: {
     // initialize values in the prescribe form
     clearForm(props.actions, props?.prefillNotes ? { notes: props.prefillNotes } : undefined);
   });
-
-  const dispatchOrderError = (errors: readonly GraphQLFormattedError[]) => {
-    const event = new CustomEvent('photon-order-error', {
-      composed: true,
-      bubbles: true,
-      detail: {
-        errors: errors
-      }
-    });
-    ref?.dispatchEvent(event);
-  };
 
   const handleAddPrescription = async () => {
     setIsLoading(true);
@@ -175,298 +164,295 @@ export const AddPrescriptionCard = (props: {
   });
 
   return (
-    <div ref={ref}>
-      <style>{photonStyles}</style>
-      <Card addChildrenDivider={true}>
-        <Text color="gray">Add Prescription</Text>
-        <div
-          class="flex flex-col"
-          on:photon-medication-selected={(e: any) => {
-            setOffCatalog(e.detail.medication);
-            props.actions.updateFormValue({
-              key: 'treatment',
-              value: e.detail.medication
-            });
+    <Card addChildrenDivider={true}>
+      <Text color="gray">Add Prescription</Text>
+      <div
+        class="flex flex-col"
+        on:photon-medication-selected={(e: any) => {
+          setOffCatalog(e.detail.medication);
+          props.actions.updateFormValue({
+            key: 'treatment',
+            value: e.detail.medication
+          });
+
+          props.draftedPrescriptionChanged();
+        }}
+      >
+        <photon-medication-search
+          label="Search for Treatment"
+          catalog-id={props.catalogId}
+          allow-off-catalog-search={props.allowOffCatalogSearch}
+          selected={props.store.treatment?.value ?? undefined}
+          invalid={props.store.treatment?.error ?? false}
+          help-text={props.store.treatment?.error}
+          off-catalog-option={offCatalog()}
+          search-text={searchText()}
+          disable-list={props.disableList}
+          on:photon-treatment-selected={(e: any) => {
+            if (e.detail.data.__typename === 'PrescriptionTemplate') {
+              repopulateForm(props.actions, {
+                ...e.detail.data,
+                notes: [e.detail.data?.notes, props.prefillNotes].filter((x) => x).join('\n\n')
+              });
+            } else {
+              props.actions.updateFormValue({
+                key: 'treatment',
+                value: e.detail.data
+              });
+            }
+
+            if (e.detail.catalogId) {
+              props.actions.updateFormValue({
+                key: 'catalogId',
+                value: e.detail.catalogId
+              });
+            }
 
             props.draftedPrescriptionChanged();
           }}
-        >
-          <photon-medication-search
-            label="Search for Treatment"
-            catalog-id={props.catalogId}
-            allow-off-catalog-search={props.allowOffCatalogSearch}
-            selected={props.store.treatment?.value ?? undefined}
-            invalid={props.store.treatment?.error ?? false}
-            help-text={props.store.treatment?.error}
-            off-catalog-option={offCatalog()}
-            search-text={searchText()}
-            disable-list={props.disableList}
-            on:photon-treatment-selected={(e: any) => {
-              if (e.detail.data.__typename === 'PrescriptionTemplate') {
-                repopulateForm(props.actions, {
-                  ...e.detail.data,
-                  notes: [e.detail.data?.notes, props.prefillNotes].filter((x) => x).join('\n\n')
-                });
-              } else {
-                props.actions.updateFormValue({
-                  key: 'treatment',
-                  value: e.detail.data
-                });
-              }
+          on:photon-treatment-unselected={() => {
+            clearForm(
+              props.actions,
+              props?.prefillNotes ? { notes: props.prefillNotes } : undefined
+            );
 
-              if (e.detail.catalogId) {
-                props.actions.updateFormValue({
-                  key: 'catalogId',
-                  value: e.detail.catalogId
-                });
-              }
+            props.draftedPrescriptionChanged();
+          }}
+          on:photon-search-text-changed={(e: any) => setSearchText(e.detail.text)}
+        />
 
-              props.draftedPrescriptionChanged();
-            }}
-            on:photon-treatment-unselected={() => {
-              clearForm(
-                props.actions,
-                props?.prefillNotes ? { notes: props.prefillNotes } : undefined
-              );
+        <ScreeningAlerts
+          /** we'll want to make sure we're only showing screening alerts that are involved with this entity */
+          screeningAlerts={props.screeningAlerts.filter(
+            (screeningAlert) =>
+              screeningAlert.involvedEntities
+                .map((involvedEntity) => involvedEntity.id)
+                .indexOf(props.store.treatment?.value?.id) >= 0
+          )}
+          owningId={props.store.treatment?.value?.id}
+        />
 
-              props.draftedPrescriptionChanged();
-            }}
-            on:photon-search-text-changed={(e: any) => setSearchText(e.detail.text)}
-          />
-
-          <ScreeningAlerts
-            /** we'll want to make sure we're only showing screening alerts that are involved with this entity */
-            screeningAlerts={props.screeningAlerts.filter(
-              (screeningAlert) =>
-                screeningAlert.involvedEntities
-                  .map((involvedEntity) => involvedEntity.id)
-                  .indexOf(props.store.treatment?.value?.id) >= 0
-            )}
-            owningId={props.store.treatment?.value?.id}
-          />
-
-          <div class="flex flex-col sm:flex-none sm:grid sm:grid-cols-2 sm:gap-4">
-            <div class="order-last sm:order-first">
-              <photon-checkbox
-                label="Dispense as written"
-                tip="This prescription will be filled generically unless this box is checked"
-                form-name="daw"
-                class="flex-grow"
-                checked={props.store.dispenseAsWritten?.value || false}
-                on:photon-checkbox-toggled={(e: any) =>
-                  props.actions.updateFormValue({
-                    key: 'dispenseAsWritten',
-                    value: e.detail.checked
-                  })
-                }
-              />
-            </div>
-          </div>
-          <div class="mt-2 sm:mt-0 sm:grid sm:grid-cols-2 sm:gap-4">
-            <div class="flex items-end gap-1 items-stretch">
-              <photon-number-input
-                class="flex-grow flex-1 w-2/5 sm:w-auto"
-                label="Quantity"
-                value={props.store.dispenseQuantity?.value ?? null}
-                required="true"
-                min={0}
-                invalid={props.store.dispenseQuantity?.error ?? false}
-                help-text={props.store.dispenseQuantity?.error}
-                on:photon-input-changed={(e: any) => {
-                  const inputValue = Number(e.detail.input);
-                  // this handles a bug on mobile where the input is cleared when the user types a decimal.
-                  // However, this introduces a bug where the input validator isn't registered. To fix this, we
-                  // add a undefined form update in the onMount function up top 🙄
-                  // https://github.com/Photon-Health/client/commit/9566daa5dea50709677c66fdceac6d2edbd43fe5
-                  if (!isNaN(inputValue) && e.detail.input !== '') {
-                    props.actions.updateFormValue({
-                      key: 'dispenseQuantity',
-                      value: inputValue
-                    });
-                  }
-                }}
-                style={{ width: '100px' }}
-              />
-              <DoseCalculator
-                open={openDoseCalculator()}
-                onClose={() => setOpenDoseCalculator(false)}
-                medicationName={props.store.treatment?.value?.name}
-                weight={props.weight}
-                weightUnit={props.weightUnit}
-                setAutocompleteValues={({ liquidDose, totalLiquid, unit, days }) => {
-                  props.actions.updateFormValue({
-                    key: 'daysSupply',
-                    value: Number(days)
-                  });
-                  props.actions.updateFormValue({
-                    key: 'dispenseQuantity',
-                    value: Number(totalLiquid)
-                  });
-                  props.actions.updateFormValue({
-                    key: 'instructions',
-                    value: `${liquidDose} ${unit} ${props.store.instructions?.value}`
-                  });
-                  if (unit === 'mL') {
-                    props.actions.updateFormValue({
-                      key: 'dispenseUnit',
-                      value: 'Milliliter'
-                    });
-                  }
-                }}
-              />
-              <div>
-                <Button
-                  variant="secondary"
-                  class="w-fit"
-                  onClick={() => setOpenDoseCalculator(true)}
-                  style={{
-                    // ya, it ain't pretty, but it works. just need it for a lil bit longer
-                    height: '40px',
-                    'margin-top': '32px'
-                  }}
-                >
-                  <Icon name="calculator" size="sm" />
-                </Button>
-                <div style={{ height: '23px' }} class="pt-1" />
-              </div>
-            </div>
-            <photon-dispense-units
-              label="Dispense Unit"
-              required="true"
-              force-label-size="true"
-              selected={props.store.dispenseUnit?.value ?? dispenseUnit()?.name}
-              invalid={props.store.dispenseUnit?.error ?? false}
-              help-text={props.store.dispenseUnit?.error}
-              on:photon-dispense-unit-selected={(e: any) => {
+        <div class="flex flex-col sm:flex-none sm:grid sm:grid-cols-2 sm:gap-4">
+          <div class="order-last sm:order-first">
+            <photon-checkbox
+              label="Dispense as written"
+              tip="This prescription will be filled generically unless this box is checked"
+              form-name="daw"
+              class="flex-grow"
+              checked={props.store.dispenseAsWritten?.value || false}
+              on:photon-checkbox-toggled={(e: any) =>
                 props.actions.updateFormValue({
-                  key: 'dispenseUnit',
-                  value: e.detail.dispenseUnit.name
-                });
-              }}
-            />
-          </div>
-          <div class="sm:grid sm:grid-cols-2 sm:gap-4">
-            <photon-number-input
-              class="flex-grow flex-shrink flex-1"
-              label="Days Supply"
-              value={props.store.daysSupply?.value ?? null}
-              invalid={props.store.daysSupply?.error ?? false}
-              help-text={props.store.daysSupply?.error}
-              required="true"
-              min={0}
-              on:photon-input-changed={(e: any) => {
-                props.actions.updateFormValue({
-                  key: 'daysSupply',
-                  value: Number(e.detail.input)
-                });
-              }}
-            />
-            <photon-number-input
-              class="flex-grow flex-shrink flex-1"
-              label="Refills"
-              value={props.store.refillsInput?.value}
-              required="true"
-              min={0}
-              max={11}
-              invalid={props.store.refillsInput?.error ?? false}
-              help-text={props.store.refillsInput?.error}
-              on:photon-input-changed={(e: any) =>
-                props.actions.updateFormValue({
-                  key: 'refillsInput',
-                  value: Number(e.detail.input)
+                  key: 'dispenseAsWritten',
+                  value: e.detail.checked
                 })
               }
             />
-          </div>
-          <photon-textarea
-            label="Patient Instructions (SIG)"
-            form-name="patient_instructions"
-            required="true"
-            placeholder="Enter patient instructions"
-            invalid={props.store.instructions?.error ?? false}
-            help-text={props.store.instructions?.error}
-            on:photon-textarea-changed={(e: any) =>
-              props.actions.updateFormValue({
-                key: 'instructions',
-                value: e.detail.value
-              })
-            }
-            value={props.store.instructions?.value}
-          />
-          <photon-textarea
-            label="Pharmacy Note"
-            placeholder="Enter pharmacy note"
-            on:photon-textarea-changed={(e: any) =>
-              props.actions.updateFormValue({
-                key: 'notes',
-                value: e.detail.value
-              })
-            }
-            value={props.store.notes?.value}
-          />
-          <div class="w-full">
-            <photon-datepicker
-              value={props.store.doNotFillBeforeDate?.value}
-              label="Do Not Fill Before"
-              invalid={props.store.doNotFillBeforeDate?.error ?? false}
-              help-text={props.store.doNotFillBeforeDate?.error}
-              min={new Date()}
-              on:photon-datepicker-selected={(e: any) =>
-                props.actions.updateFormValue({
-                  key: 'doNotFillBeforeDate',
-                  value: e.detail.date
-                })
-              }
-            />
-          </div>
-          <div class="flex flex-col xs:flex-row gap-2">
-            <Show when={!props.hideAddToTemplates}>
-              <photon-checkbox
-                label="Add To Personal Templates"
-                form-name="addToTemplates"
-                checked={props.store.addToTemplates?.value || false}
-                on:photon-checkbox-toggled={(e: any) => {
-                  props.actions.updateFormValue({
-                    key: 'addToTemplates',
-                    value: e.detail.checked
-                  });
-                }}
-              />
-            </Show>
-            <Show when={props.store.addToTemplates?.value ?? false}>
-              <photon-text-input
-                class="flex-grow flex-shrink flex-1"
-                label="Template Name"
-                value={props.store.templateName?.value ?? ''}
-                invalid={props.store.templateName?.error ?? false}
-                help-text={props.store.templateName?.error}
-                on:photon-input-changed={(e: any) =>
-                  props.actions.updateFormValue({
-                    key: 'templateName',
-                    value: e.detail.input
-                  })
-                }
-              />
-            </Show>
-            <div class="flex flex-grow justify-end">
-              <Button
-                class="w-full xs:!w-auto h-fit"
-                size="lg"
-                onClick={() => {
-                  if (!isLoading()) {
-                    handleAddPrescription();
-                  }
-                }}
-                loading={isLoading()}
-                variant="secondary"
-                color="blue"
-              >
-                Add to drafts
-              </Button>
-            </div>
           </div>
         </div>
-      </Card>
-    </div>
+        <div class="mt-2 sm:mt-0 sm:grid sm:grid-cols-2 sm:gap-4">
+          <div class="flex items-end gap-1 items-stretch">
+            <photon-number-input
+              class="flex-grow flex-1 w-2/5 sm:w-auto"
+              label="Quantity"
+              value={props.store.dispenseQuantity?.value ?? null}
+              required="true"
+              min={0}
+              invalid={props.store.dispenseQuantity?.error ?? false}
+              help-text={props.store.dispenseQuantity?.error}
+              on:photon-input-changed={(e: any) => {
+                const inputValue = Number(e.detail.input);
+                // this handles a bug on mobile where the input is cleared when the user types a decimal.
+                // However, this introduces a bug where the input validator isn't registered. To fix this, we
+                // add a undefined form update in the onMount function up top 🙄
+                // https://github.com/Photon-Health/client/commit/9566daa5dea50709677c66fdceac6d2edbd43fe5
+                if (!isNaN(inputValue) && e.detail.input !== '') {
+                  props.actions.updateFormValue({
+                    key: 'dispenseQuantity',
+                    value: inputValue
+                  });
+                }
+              }}
+              style={{ width: '100px' }}
+            />
+            <DoseCalculator
+              open={openDoseCalculator()}
+              onClose={() => setOpenDoseCalculator(false)}
+              medicationName={props.store.treatment?.value?.name}
+              weight={props.weight}
+              weightUnit={props.weightUnit}
+              setAutocompleteValues={({ liquidDose, totalLiquid, unit, days }) => {
+                props.actions.updateFormValue({
+                  key: 'daysSupply',
+                  value: Number(days)
+                });
+                props.actions.updateFormValue({
+                  key: 'dispenseQuantity',
+                  value: Number(totalLiquid)
+                });
+                props.actions.updateFormValue({
+                  key: 'instructions',
+                  value: `${liquidDose} ${unit} ${props.store.instructions?.value}`
+                });
+                if (unit === 'mL') {
+                  props.actions.updateFormValue({
+                    key: 'dispenseUnit',
+                    value: 'Milliliter'
+                  });
+                }
+              }}
+            />
+            <div>
+              <Button
+                variant="secondary"
+                class="w-fit"
+                onClick={() => setOpenDoseCalculator(true)}
+                style={{
+                  // ya, it ain't pretty, but it works. just need it for a lil bit longer
+                  height: '40px',
+                  'margin-top': '32px'
+                }}
+              >
+                <Icon name="calculator" size="sm" />
+              </Button>
+              <div style={{ height: '23px' }} class="pt-1" />
+            </div>
+          </div>
+          <photon-dispense-units
+            label="Dispense Unit"
+            required="true"
+            force-label-size="true"
+            selected={props.store.dispenseUnit?.value ?? dispenseUnit()?.name}
+            invalid={props.store.dispenseUnit?.error ?? false}
+            help-text={props.store.dispenseUnit?.error}
+            on:photon-dispense-unit-selected={(e: any) => {
+              props.actions.updateFormValue({
+                key: 'dispenseUnit',
+                value: e.detail.dispenseUnit.name
+              });
+            }}
+          />
+        </div>
+        <div class="sm:grid sm:grid-cols-2 sm:gap-4">
+          <photon-number-input
+            class="flex-grow flex-shrink flex-1"
+            label="Days Supply"
+            value={props.store.daysSupply?.value ?? null}
+            invalid={props.store.daysSupply?.error ?? false}
+            help-text={props.store.daysSupply?.error}
+            required="true"
+            min={0}
+            on:photon-input-changed={(e: any) => {
+              props.actions.updateFormValue({
+                key: 'daysSupply',
+                value: Number(e.detail.input)
+              });
+            }}
+          />
+          <photon-number-input
+            class="flex-grow flex-shrink flex-1"
+            label="Refills"
+            value={props.store.refillsInput?.value}
+            required="true"
+            min={0}
+            max={11}
+            invalid={props.store.refillsInput?.error ?? false}
+            help-text={props.store.refillsInput?.error}
+            on:photon-input-changed={(e: any) =>
+              props.actions.updateFormValue({
+                key: 'refillsInput',
+                value: Number(e.detail.input)
+              })
+            }
+          />
+        </div>
+        <photon-textarea
+          label="Patient Instructions (SIG)"
+          form-name="patient_instructions"
+          required="true"
+          placeholder="Enter patient instructions"
+          invalid={props.store.instructions?.error ?? false}
+          help-text={props.store.instructions?.error}
+          on:photon-textarea-changed={(e: any) =>
+            props.actions.updateFormValue({
+              key: 'instructions',
+              value: e.detail.value
+            })
+          }
+          value={props.store.instructions?.value}
+        />
+        <photon-textarea
+          label="Pharmacy Note"
+          placeholder="Enter pharmacy note"
+          on:photon-textarea-changed={(e: any) =>
+            props.actions.updateFormValue({
+              key: 'notes',
+              value: e.detail.value
+            })
+          }
+          value={props.store.notes?.value}
+        />
+        <div class="w-full">
+          <photon-datepicker
+            value={props.store.doNotFillBeforeDate?.value}
+            label="Do Not Fill Before"
+            invalid={props.store.doNotFillBeforeDate?.error ?? false}
+            help-text={props.store.doNotFillBeforeDate?.error}
+            min={new Date()}
+            on:photon-datepicker-selected={(e: any) =>
+              props.actions.updateFormValue({
+                key: 'doNotFillBeforeDate',
+                value: e.detail.date
+              })
+            }
+          />
+        </div>
+        <div class="flex flex-col xs:flex-row gap-2">
+          <Show when={!props.hideAddToTemplates}>
+            <photon-checkbox
+              label="Add To Personal Templates"
+              form-name="addToTemplates"
+              checked={props.store.addToTemplates?.value || false}
+              on:photon-checkbox-toggled={(e: any) => {
+                props.actions.updateFormValue({
+                  key: 'addToTemplates',
+                  value: e.detail.checked
+                });
+              }}
+            />
+          </Show>
+          <Show when={props.store.addToTemplates?.value ?? false}>
+            <photon-text-input
+              class="flex-grow flex-shrink flex-1"
+              label="Template Name"
+              value={props.store.templateName?.value ?? ''}
+              invalid={props.store.templateName?.error ?? false}
+              help-text={props.store.templateName?.error}
+              on:photon-input-changed={(e: any) =>
+                props.actions.updateFormValue({
+                  key: 'templateName',
+                  value: e.detail.input
+                })
+              }
+            />
+          </Show>
+          <div class="flex flex-grow justify-end">
+            <Button
+              class="w-full xs:!w-auto h-fit"
+              size="lg"
+              onClick={() => {
+                if (!isLoading()) {
+                  handleAddPrescription();
+                }
+              }}
+              loading={isLoading()}
+              variant="secondary"
+              color="blue"
+            >
+              Add to drafts
+            </Button>
+          </div>
+        </div>
+      </div>
+    </Card>
   );
 };
